requirements.txt
# Core deep learning
torch>=2.1.0
torchvision>=0.16.0

# Implicit dynamics / DEQ
torchdeq>=0.1.0

# Scientific computing
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualisation
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: dimensionality reduction (for latent trajectory plots)
umap-learn>=0.5.3

# Biomedical time-series data
wfdb>=4.1.0

# Config and utilities
pyyaml>=6.0
tqdm>=4.65.0
einops>=0.7.0

# Optional: experiment tracking (not required for evaluation)
# wandb>=0.16.0

# Testing
pytest>=7.4.0